Move the test to integration
