+++ integration/format/dumplib.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package dump_tree
+package format
 
 import (
 	"bytes"
